+++ cmd/exchaind/mpt/mpt_viewer.go
@@ -8,7 +8,7 @@ import (
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/trie"
 	"github.com/okex/exchain/libs/cosmos-sdk/server"
-	"github.com/okex/exchain/libs/mpt"
+	"github.com/okex/exchain/libs/cosmos-sdk/store/mpt"
 	"github.com/spf13/cobra"
 )
