+++ libs/cosmos-sdk/store/mpt/types/wrap_rocksdb_batch.go
