+++ libs/cosmos-sdk/x/distribution/keeper/test_common.go
@@ -1,8 +1,9 @@
 package keeper
 
 import (
-	"github.com/okex/exchain/libs/mpt"
 	"testing"
+
+	"github.com/okex/exchain/libs/cosmos-sdk/store/mpt"
 
 	"github.com/stretchr/testify/require"
