+++ libs/cosmos-sdk/x/gov/keeper/test_common.go
@@ -6,8 +6,9 @@ package keeper // noalias
 import (
 	"bytes"
 	"encoding/hex"
-	"github.com/okex/exchain/libs/mpt"
 	"testing"
+
+	"github.com/okex/exchain/libs/cosmos-sdk/store/mpt"
 
 	"github.com/stretchr/testify/require"
